+++ example/android/app/src/main/kotlin/com/sayegh/move_to_background_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.sayegh.move_to_background_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}